Work in progress heater and thermistor code. Update and simplify spectral code.

+++ src/util/hardware_adc.hpp
@@ -27,7 +27,6 @@ namespace mrover {
 		}
 
 		void update() {
-
 			HAL_ADC_Start_DMA(m_hadc, reinterpret_cast<uint32_t*>(m_values.data()), m_channels);
 //			HAL_ADC_Start_DMA(m_hadc, reinterpret_cast<uint32_t*>(m_values.data()), m_channels / (sizeof(uint32_t) / sizeof(uint16_t)));
 		}
@@ -35,7 +34,7 @@ namespace mrover {
 	private:
 		ADC_HandleTypeDef* m_hadc;
 		uint8_t m_channels;
-		std::vector<uint16_t> m_values;
+		std::vector<uint32_t> m_values;
 	};
 
 } // namespace mrover
